exclude non-transactional tables
