+++ ql/src/java/org/apache/hadoop/hive/ql/io/IOContext.java
@@ -20,6 +20,8 @@ package org.apache.hadoop.hive.ql.io;
 
 import org.apache.hadoop.fs.Path;
 
+import static org.apache.hadoop.hive.ql.io.AcidUtils.parseSplitPath;
+
 /**
  * IOContext basically contains the position information of the current
  * key/value. For blockCompressed files, isBlockPointer should return true,
@@ -179,33 +181,6 @@ public class IOContext {
 
   public void parseRecordIdentifier() {
     this.ri = parseSplitPath(inputPath);
-  }
-
-  public static RecordIdentifier parseSplitPath(Path splitPath) {
-    RecordIdentifier recordIdentifier = new RecordIdentifier();
-    Path parent = splitPath.getParent();
-    if (parent == null) {
-      return null;
-    }
-    boolean isBase = parent.getName().startsWith(AcidUtils.BASE_PREFIX);
-    boolean isDelta = parent.getName().startsWith(AcidUtils.DELTA_PREFIX)
-        || parent.getName().startsWith(AcidUtils.DELETE_DELTA_PREFIX);
-    if (isBase || isDelta) {
-      if (isBase) {
-        recordIdentifier.setValues(
-            AcidUtils.ParsedBaseLight.parseBase(parent).getWriteId(),
-            AcidUtils.parseBucketId(splitPath),
-            0);
-      } else {
-        AcidUtils.ParsedDeltaLight pd = AcidUtils.ParsedDeltaLight.parse(parent);
-        recordIdentifier.setValues(
-            pd.getMinWriteId(),
-            AcidUtils.parseBucketId(splitPath),
-            0);
-      }
-    }
-
-    return recordIdentifier;
   }
 
   public boolean isDeletedRecord() {
